Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

I am not a great fan of such specific guidelines. I think the transcend the "ease of understanding" and go into "squashing individual personality", but that is a different discussion, that may be better held elsewhere.

Meanwhile, I noticed that the third file also has a long line problem, and I'll fix it and resubmit.

Patch-set: 2
Label: Code-Review=-1