Update patch set 3
Patch Set 3:

You don't have to be a fan... you just have to accept it.  We all write code that looks the same because it's then easier to read and maintain.

Not everybody's going to be happy (few are) because the rules are all very subjective, but everybody's happy at least that they can read code in another component and understand it quickly because the style & naming & other conventions are the same.

Patch-set: 3
Reviewer: Gerrit User 1000269 <1000269@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0
Label: Verified=0