Update patch set 2
Patch Set 2: Do not submit

I actually am going to say I prefer keeping weak references.  Reasons:

(1) I am told that the current implementation of soft references is to release half of them, randomly selected, when a gc happens.  Thus I don't see this really gaining a whole lot, and introducing yet more randomness in behavior that I wouldn't like to have.

(2) The main purpose of this cache is to get rid of *duplicate* objects.  That is, if you have two views in a layout both with the same background drawable, we don't want to load the drawable twice, but just re-use the one we already have.  This is what weak references give us.

(3) The drawables in the cache can represent a significant amount of data (by being a bitmap), so letting them linger outside of the application's control is something I would like to avoid.  This would just beg for us to then have an API for an app to say "don't hold a soft reference on this object."  Instead, if an application has certain resource objects that it doesn't want to let be expunged from the cache, it can just explicitly hold its own reference on the object.

(4) The only way I can see this improving performance when switching from landscape to portrait is if a GC happens in the middle of that after all references to the old activity have have gone away and the new one is not yet constructed.  If this is an issue (maybe the new concurrent GC makes this more likely to happen?), we could mitigate that by holding the references longer.  Fwiw, I do see at a rotation more resources being loaded than I would like, but from what I can see most of these should still be in the cache even if it is using weak references.  So I would like to look specifically at what is going on there instead of just changing these to soft references.

Patch-set: 2
Label: Code-Review=-2