stagefright IOMX: Provide a writeable copy of the setParameters argument
The qcom OMX component in Nexus One (the binary distributed with froyo
at least) will try to write into the parameter data, which is allowed
according to the OpenMAX IL spec, where the parameter isn't a const
pointer.

When setParameter is called over the binder IOMX interface with this
OMX component, it leads to a segfault since the parameter data is
read only.

Change-Id: I0f6bcdf99a9ef1234cb27b741da4294f39df9b11