Update patch set 5
Patch Set 5:

From author:


The reason for not including mSimRecords.setVoiceCallForwardingFlag(1, enabled) in setCallForwardingPreference() is that it will cause a problem at startup if the SIM card has been changed. If the SIM card has been changed the preference is set to false. We don't want to set the CF field to false on the SIM card just because it has been changed, since we don't actually know the current CF status at this point. 

As for putting mSimRecords.getVoiceCallForwardingFlag() in getCallForwardingPreference(). The method getCallForwardingPreference() is only used next to mSimRecords.getVoiceCallForwardingFlag one time, in updateCallForwardingIndicator(). In all other cases the method it is used when we only want to get the value of the preference not the value stored in SimRecords. Given the above, I believe it's better to leave the code as-is.

Patch-set: 5