Update patch set 4
Patch Set 4: Looks good to me, but someone else must approve

Should be on master branch now.
    
I have tested it before on the froyo branch in the emulator via eclipse, however often I cannot get the master branch to build in eclipse, or before that in the terminal via make even. Any advice? I presume it is just the state of the master branch that it does not always compile. This resubmit to the master branch has not been tested an sich, but since the code is identical and the change is localized, the risk is minimal.

Patch-set: 4
Reviewer: Gerrit User 1003335 <1003335@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1