libstagefright: fix live RTSP streaming on some servers
'npt=0-' is also a valid range to describe an "infinite" session, and
it's used by some popular RTSP servers in mobile environments (Envivio's,
for example). Assuming that infinite is always 'now-' was causing session
SETUPs to fail (and thus breaking the "Mobile TV"-type services).

Also, drop RTP video packets that are out of sequence. Trying to
reorder these was causing a *very* long time in starting the streams.

Change-Id: Id5d5ccfae2e70d9b0a0dfb8f120bf1a9b6d61b5f