Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

With a previous version of this change, I was a little concerned that changing the values of those constants might impact existing entries in the MMS database and how they are handled by the app and the content provider. I believe that Tom said that the earlier change was fine. I'll double check if I can see any problems that might arise with changing these values, but it is definitely better to start using the correct values than the incorrect ones. So, LGTM.

Patch-set: 2
Reviewer: Gerrit User 1003544 <1003544@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1
Label: Verified=0