Add support for dynamic resource redirection.
This change supports resource redirection by either system or user
installed theme packages which affect the final asset delivered to
applications.  Theme packages can redirect assets from either the
framework or other third party packages.  For instance, assets in the
launcher or in the contacts app can be themed by any third party theme.

Theme packages differ from regular application packages in their
manifest structure with an additional meta data XML format that states
the explicit resource redirections to be looked up by name at runtime.
Because theme packages coexist in the same AssetManager as regular
packages they are compiled in a different package scope (currently 0x0a)
to avoid conflicts with the framework (0x01) and applications (0x7f).

Redirections are maintained in a new system service whose sole purpose
is to cache the resource id mappings in memory and provide it to
individual app processes on demand.  This cache speeds up asset manager
construction when theming is required but is not persisted across
reboots as the cost to construct each individual mapping is relatively
low.

Themes can be swapped dynamically at runtime without the need for a
device restart.  This is possible by attaching the current user theme to
the Configuration object which, upon an unhandled change, will trigger
the usual activity onDestroy/onCreate cycle.  Additionally, the ResTable
class was modified to support dynamic updates of the attached asset
paths (the old theme is removed and the newly applied theme is added
dynamically).  Updates to some special areas of the system UI must be
handled individually like the status bar service as it does not have a
conventional activity lifecycle.

Finally, this patch is supported by an application suite (with a
compatible license) that manages the installed packages and allows the
user to apply their preferred theme.  Though it is the primary intention
of this feature to support user controllable themes it also enables
device manufacturers to "skin" the default Android styles by simply
providing a default system theme and choosing not to offer the
application suite.

More information can be found at http://tmobile-themes.googlecode.com/.

Change-Id: I6be0b58b77ea6236c812eb52d86e76c69186b00b