Support for run-time resource overlay.
This patch introduces support for package resources to be overlaid in
run-time. When a package is loaded, the presence of a special mirror
package containing resources only is checked. If present, the overlay
package is loaded alongside the regular package. The latter's resource
table is amended with values from the overlay package; file operations
are redirected to resources in the overlay package should it hold a
mirror version of the requested resource.

To achieve consistent resource IDs in overlay package and its target,
the implementation requires overlay packages contain a mapping table
between target and overlay resource IDs. The map is constructed during
run-time, during apk installation, and cached similarly to .odex files.

Run-time resource overlay may be used to
- decouple customization and application development
- enable rapid prototyping and debugging

See README for further details.

Change-Id: If26ee7754813004a96c043dba37fbe99fa3919db