Checking strings prior sending them to JNI.
The MediaScannerClient was sending raw strings from external
content directly to JNI. When JNI checking is enabled this
causes crashes when there is invalid data on the SD card.

This change adds checks for non-UTF8 characters that we
cannot find any valid encoding for is replaced with "???"
before sending strings to JNI. The check itself is
copied from the Dalvik JNI tests.

Change-Id: Ica51db9c3a93e953bf87542acdf7ca733c9a6fbe