Added bidi layout support to the view framework
In order to minimize the effort for app developers when
adapting application layouts for languages with
alphabets read from right-to-left the view framework
has been extended to cope with directionality
information. A layout does by default not have a
directionality, but when specified to be viewed from
left-to-right the position of contained views will
automatically be "mirrored" when the phone laguage
alphabet is read from right-to-left. The framework
widgets have also been adapted to be drawn in "mirrored"
state in this situation.

Change-Id: I491eded164bd6dd6c2ba5573fb377ab8c763f6d7