Add locking to captureScreenImplLocked()
If an unlucky application tries to capture the screen at the wrong
time, it can get a corrupted screen image.

This patch prevents this problem by locking the layers while drawing
the screenshot, so that they can't redraw at the same time and the
caller always gets an atomic snapshot of the screen image.

Change-Id: I43126b9b4be2c03196d21bf28e938512243b2b0e