Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Thank you for your submission but this is not how we want to support RTL layouts. We want to do it more extensively with the notion of before and after gravity (in addition to left and right that should remain left and right in RTL mode.) We also want the notion of context, etc.
We have an internal design defining such an implementation as well as patches. I'll let you know more about this as soon as I can.

Patch-set: 1