Fix for issue related to expandable list.
The user experience is that of a lost touch event. Typically the user
first expands an item successfully and then attempt to collapse the item,
only to fail on the first attempt. A more technical description would be
that the first touch event will trigger a scrolling procedure, which in
a certain case will fail because of inconsistent parameter (scrolling
distance equal to or less then zero). This fix do not change existing code, but
rather adds handling of this special state. Reproduce 10/10, by first
expanding a list item, then expanding one more list item further down
in the list and if all items in the fully expanded list is visible on the
screen (scrolling is not necessary).

Change-Id: Id083e9d8a1085845d1347069452b156ae2c33e03