Improve shutdown speed by disabling radio/bt early
Improve shutdown speed by disabling radio/bt as the first thing.
Before buzzing we still check radio/bt status to be sure they are disabled.
If not we wait as much as before but with a more often (100ms) check on
status compared to previous 500ms. This way other shutdown operations can
be done in paralel, reducing total time to take the system down. This
shaves a few seconds of the total shutdown time.

Change-Id: Icd5eb5083ec64b4c57d32413a258b499a0711de7