Update patch set 3
Patch Set 3:

It would seem to me as if the existing code already handles the case:

Cache-Control: max-age=200000, no-cache

as it then parses max-age=200000 and sets ret.expires. After this, the no-cache directive is found, and then ret.expires is set to 0.

The new code introduced handles the case:

Cache-Control: no-cache, max-age=200000

where previously, no-cache was found, and ret.expires was set to 0, and after that, the max-age header still was parsed.

Patch-set: 3
Reviewer: Gerrit User 1003744 <1003744@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0