Runtime inclusion of services in the SystemServer with overlays
Its a common problem with developers writing applications
that run in separate processes to host system services, more
than the ones that Android currently has defined. This commit
introduces a way to include more services to be started inside
the SystemServer, with the services to be started that are defined
in the resource framework. Having the services defined in the
resource framework enable device vendors to include their
own services with the overlay mechanism.

Change-Id: I8c33c92f38e4a5dd095d1138847c7d2bab6b0769