Enable possibility to register for ACTION_HEADSET_PLUG statically
Currently, its not possible to register for ACTION_HEADSET_PLUG in
the manifest of Android applications, and the intent is only available
when the application is already running. This leads to odd use-cases
where applications needs to start on BOOT_COMPLETEd, to register for
the intent, and then stay alive. Instead of forcing applications
to do this, this commit enables them to listen for this intent
in AndroidManifest.xml.

Change-Id: Ibee7ab5d47bd1972d0b8cf98de735fa030e3e7d0