Update patch set 5
Patch Set 5:

This change will not fix the issue if the SIM card is removed and inserted in another phone. Since the SIM card doesn't have any information stored for the call forwarding the new phone would not show the icon.

Would it be better to call getCallForwardingOption method on phone boot up to query the call forwarding information from the network instead?

Patch-set: 5