Update patch set 2
Patch Set 2:

This is the basic mechanism for controlling which services to start, it does not actually change anything at this point. If this one is ok then we could move some of the current ones out into the config, e.g. the dockobserver for devices that cannot be docked. another useful case is to be able to use a different headset observer for devices with a totally different behaviour of the headset drivers.

Patch-set: 2