Update patch set 2
Patch Set 2: (1 inline comment)

I will update both the commit message, and the comment in the config file. 

As Johan explained, this could be used to both add and remove services in this manner, to optimize for the current product, but I agree, it deserves a better commit message. Will do that tomorrow morning.

Patch-set: 2