Enable notifications during an ongoing call
If a phone call is ongoing, the alert will be allowed but only
using a special incall notification tone and no vibration. The tone will
be played on the voice call audio stream.

The requirement behind this is to allow audible SMS/MMS notifications
during phone calls, something that is requested by a number of
European operators.

Change-Id: I89326cb0768cf8b8ffda979d038e4c0788a8b43a