SF: Get state lock before reading transact flags
If the state lock is not held when reading the transaction flags there
is a chance that the state is changed between the reading of the
transaction flags and the acquiring of the state lock, if that happens
the transaction handler could miss state changes. Example:

1. The transaction handler reads the transaction flags, only
"traversal" is set.
2. A layer is removed and the transaction flags are updated, setting
"transaction".
3. The transaction handler acquires the state lock.
4. The transaction handler does its stuff. It will not notice the
removed layer as its version of the transaction flags does not indicate
that any layer has been removed.
5. The transaction handler committs current to draw. Beyond this point
the transaction handler has no way of noticing the removed layer i.e.
it missed it.

This can probably lead to all kinds of problems but the only only one
investigated is the one described in the example above. The result of
the transaction handler missing a layer removal is that the Layer
remains forever in the purgatory list which is bad as the layer still
holds references to its buffers.

Change-Id: I41ee8b3c9134545bddeb2985dcbd587c8fa43e8c
Signed-off-by: Christian Bejram <christian.bejram@stericsson.com>