Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Ahh, I see the problem.

The fix isn't quite what we want though.  There could be multiple devices that support the same switches (or keys!).  Ideally we want the getState function to return UNKNOWN if none of the devices support the switch, DOWN or VIRTUAL if at least one of the devices indicates the switch is down, and UP if all of the devices that support the switch indicate that it is up.

Patch-set: 1
Label: Code-Review=-1
Label: Verified=0