Update patch set 5
Patch Set 5:

Do you really need to have stuff running in the system process?  I am wanting to get stuff *out* of it, not make it easier to throw more and more stuff in it.  Generally for things that aren't part of the core service (such as the telephony stack) they can be implemented as a separate .apk with android:persistent="true" so they are always running.  Oh and like in GB all of the system UI elements have been moved out of the system process into their own "SystemUI" .apk/process.

I'd like to understand why being able to dynamically add stuff to the system process like this is needed instead of running as a separate .apk.

At any rate, I don't want such a mechanism done like this.  It should be starting services that are configured to run in the system process, not just creating and running classes.

Patch-set: 5