Added dependency injection framework
This framework is aimed at giving the developer an API for
dependency injection which is simple to use and understand, where
as little time as possible is spent configuring the framework
while still providing all of the essential features. It is also
specifically optimized for use on Android and besides performance
tuning it incorporates transparent access to Android resources.

The API is currently hidden from public access using @hide.
See JavaDoc at core/java/android/inject/DuctTape.java for more
information on how the framework is used.

Change-Id: I3969bb33a6653a3b8817ff7c1e7716ec9d57fff4