Send out failiure reason when connect to an APN fails
When a connection to an APN failed, the platform will not send
the failure reason out. This is not convenient for the upper
applications, as some application may want to do show the user
an error dialog or make some decision based on the failure
reason. Some mobile carriers have requirements concerning this.

Fixed by sending out the failure reason.

Change-Id: Ib5ef1dd3c1f180298ae35e43b95ed5b73d3b020e