Allow wifi (client mode) in parallel to mobile data network
This patch allows an application to communicate to a wifi network (as client mode) and to a mobile data network (Internet) in parallel. This enables applications to communicate with services on a local wifi network (e.g. cars, amplifiers, hifi-systems) and the internet at the same time, without having to explicitly reconfigure any devices on the local wifi network. In contrast to tethering, this patch does not allow devices on the local wifi network to access the mobile data network of the android device.

description
- The user configures in the wifi settings for a specific network that internet access via this network should not be used. By default this option is not activated.
- After a connection to the wifi network is established, both radios (wifi and mobile) remain active.
- The default internet gateway as well as the DNS servers are taken from the mobile data network, so that all internet traffic from applications on the android device will be routed through this interface. DNS servers from the wifi network are not used in this case.
- Applications on the android device can now communicate with local wifi nodes and the Internet simultaneously

actors
- Wifi network without Internet access
- Mobile network like 3G, EDGE
- Device and application

example
A media application wants to receive web radio content from the internet and redirect it to a home entertainment system located on a local wifi network.

implementation details
- A checkbox was added to the wifi settings dialog
- NetworkStateTracker was modified to include the new function useAsDefaultGw(). The default implementation is true, but the WifiStateTracker overrides this to get the value corresponding to the settings
- Depending on the wifi settings:
  - the teardown of the old data connection is prevented
  - default routes are removed after successful dhcp-resolution
  - dns servers are not added

Change-Id: I7c6b093b4f2dd1e7a46c0572b422dc67387c078c

Signed-off-by: Michael Fitzner <michael.fitzner@bmw-carit.de>