Introduce an alternative API for AsyncTask.
AsyncTask makes it difficult to properly decouple it from an implicit
outerclass Activity reference.  This is especially problematic when an
AsyncTask is to be "recovered", and in this way resumed, upon
orientation change or simply abandoned if the user presses the back
button.  The detachable pattern applied here allows us instead to divide
the problem into two classes: the DetachableAsyncTask which performs the
background work, and the TaskCallbacks which respond to main thread
callbacks.  The latter class is expected to hold the implicit outerclass
reference but that can be attached, reattached, or unattached from the
former as needed.

Change-Id: I10029d41a2903e3eeddc57c2b2f9e7a82c4490c6