Update patch set 1
Patch Set 1: Do not submit

Thanks for your contribution.  We would prefer to not add this kind of utility code to the platform (where we have to support and maintain it forever more) unless it is being used extensively by applications or is needed for the platform itself.  I think you will be much happier with supplying this code to developers as a static library -- that way they can start using it immediately (don't need to wait until the minimum platform version they target to support it) and can immediately update to any newer versions you have with bug fixes or improvements.

Also, the new fragment APIs I think mitigate a lot of the need for this, with the ability to retain a Fragment instance across Activity instances.  Basically onRetainNonConfigurationInstance() is deprecated and can be fully replaced by fragments.  And this can be used on older versions of the platform down to 1.6 through the helper static library.

Patch-set: 1