Update patch set 1
Patch Set 1:

That seems fair.  Is there any interest in proceeding with my submission of an AsyncTask sample in ApiDemos?  I searched around and there are no complete examples of this class in the current samples however it is widely used by the community and, in many cases, used incorrectly.  I would like to include my DetachableAsyncTask class along with this demo for others to either reference or copy.

Patch-set: 1