Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Pascal,
the internal branch has reverted to an uncompressed format. The team thinks the compressed complexity is not needed.

It might be great if you have access to the internal branch and can prepare a change based on latest code

Patch-set: 1