Update patch set 2
Patch Set 2:

This latest patch set is a complete rewrite where we don't do simple mirroring, and are instead following the before/after design. Also, directionality is now explicitly telling the view exactly in which direction the contents should flow, as opposed to the old solution where it was just meta-data used for deciding whether to mirror or not.

Patch-set: 2