Update patch set 1
Patch Set 1:

I am arguing against changing these numbers.  These were determined when screen size support was added in 1.6, and have been a contract with apps since then.  This is a core part of our promise to developers about backwards compatibility -- applications written prior to 1.5 were not expected to work with anything besides a 320x480 mdpi screen, and so as our backwards compatibility support we make sure they do not run in an environment where they may break because of a change from that.  We did a lot of testing of applications around this, and many of them do break when given a larger screen.  We found that we could get away reasonably well with making the screen a little longer in aspect ratio for wider screen devices, but that is it.

Please please do NOT feel free to just change things like this in the platform.  A consistent platform for developers is a FUNDAMENTAL ASPECT of the value of Android.  By changing things like this, you cause your devices to behave slightly different for developers than other devices.  This is BAD for developers.  You should never do this.  The platform as it is implemented and behaves is extremely valuable, and you HARM that value for everyone by making such changes.

If changes like there were to be made to the platform, they must be made as part of an official platform release, to have consistent behavior across devices.  However, as I said, we do not want to change these values because they are important for our contract with developers for backwards compatibility.  Developers who wrote apps that target pre-1.6 devices expect their apps to run on a screen with a certain size, which this code is protecting against.  Developers who wrote applications against 1.6 or later and/or explicitly state they are resizeable can be given the extra screen space.  That is our contract with developers.  We are not breaking it.

As far as qHD displays, the comparison with larger screen devices is simplistic.  We have deliberately decided to put densities in to buckets.  This means it is NOT our goal to have the UI size stay exactly the same.  Instead, exact UI size will fluctuate between devices depending on how what the manufacturer chooses for the screen.  A device like the EVO has selected to have a larger screen at the same resolution as other devices with smaller screens; this means the UI is generally larger, and that is what some people may like.  If you are using a higher resolution screen like qHD on a smaller screen that you are going to be on the upper end of the density bucket, thus have a relatively smaller UI size.  This is simply what you have decided to do.

I think for Android that qHD makes more sense for larger screens, which otherwise tend to have a relatively lower density.  We aren't trying to make every possible screen combination work equally well; we have a middle-ground between flexibility for hardware decisions and how much complexity there is for developers and the platform.  That means you need to be thinking about these things when designing hardware and how those decisions will interact with the way the platform adjusts to the hardware.

As far as qHD falling into the same category as WVGA -- no, it doesn't, because clearly there *is* more space that the application has to put its UI in, so fundamental differences in how they interact with the screen.  It just isn't as big a difference as going up a whole size category.  It is something that the application needs to resize itself correctly to account for, which is why they need to say they are resizeable.

Patch-set: 1