getInstalledPackages does not fit in binder heap
Calling getInstalledPackages with a lot of installed
packages, can result in a list that is to large to
fit in the binder heap. When that happens the calling
side will get an empty list as return value.

This change divides the call into multiple calls
with the extra information being retrieved in a
separate call for each package.

Change-Id: I8e0c5f674bf2098adcff6d40893f94162961031f