Update patch set 2
Patch Set 2: Do not submit

Overall, I think we do not have to submit this.
It's too overdoing for the purpose and we can't estimate the side effect.

And you must use "git commit -a --amend" when you update this CL. (Do not abandon and re-upload. It's very confusing.)

Related CLs here.
https://review.source.android.com/23646
https://review.source.android.com/23647
https://review.source.android.com/23816

Patch-set: 2