Crash when starting in safe mode
The phone would crash when starting it up in safe mode if there
were two or more 3'rd party widget added to Home screen where
both widgets came from the same package.

While going through the list of installed widget providers at
start-up, unused providers are removed from this list. When starting
in safe mode, all 3rd party installed widgets are removed. Since we
were removing items from the same list we access, there was
a risk of getting an IndexOutOfBoundsException.

The fix is to create a copy of list before we traverse it.

Change-Id: I9503e3220320de0415560dd6727f3edc90d920ef