Update patch set 1
Patch Set 1:

As I commented on one of the dependent patches:

Given that the internal and public trees currently differ in trailing whitespace in these files, any patch that touches those areas will produce merge conflicts in either of the trees, until the this patch is merged.

As Conley already commented, a git diff -w (or git show -w) will show you that the patch indeed is a pure whitespace patch. If you still don't trust it without reading it line by line yourself, I can provide a shellscript that you can read through, that does the cleanup, so that you can verify that running the script produces identical result to my patch.

Patch-set: 1