stagefright: Remove unused and dubious typedefs
These typedefs aren't used anywhere, and their description
doesn't really match their actual definition. The VO_TCHAR
type is even defined as pointer for one configuration and
as non-pointer for another configuration.

This gets rid of the last use of the LINUX define in the
VO codecs (except for the amrwbenc sample application).
Within the codecs, the use of this define is unclear - does it
indicate code specific to android but not normal linux, linux
but not other unixes, unix but not windows, or gcc but not
other compilers, etc.

Originally, prior to cleanup, the define has meant many
different of these - these uses has now been cleaned up to
use the proper feature detection macros (e.g. looking for
defines automatically set by the compiler, instead of
relying on manually setting this define).

Change-Id: I1a675d653d6d305f558aa7f74c7255858dc62758