Update patch set 2
Patch Set 2:

The intent of this patch is that the LINUX define can be dropped, once this and a bunch of the other portability cleanup patches are merged, simplifying compiling of this codec across different platforms. I guess that wasn't too apparent.

But yes, at the moment this patch definitely is a low priority patch, compared to the other ones.

Patch-set: 2