Added support for multiple raw asset directories (-A option) to aapt.
In order to enable raw assets from libraries to be compiled into main
application, aapt tool has to support multiple -A arguments.
This patch adds such capability. Multiple -A are handled the same way
as multiple -S arguments. In other words, if specific asset file is
found in several -A folders, the one that appear first in
command line wins.

Change-Id: I4192d9788c4616005e4b40a79cc9f3f42d053791
Signed-off-by: Ievgenii Nazaruk <ievgenii.nazaruk@gmail.com>