libstagefright: Use pmem input buffer in extractor media track to avoid copy
Extractor uses a local MediaBuffer to read parsed input data and
the same is copied into a input pmem buffer in OMXCodec. To avoid
the copy and to protect the input video buffer, the pmem buffer pointer
for video track is passed to the Extractor which uses it to read the
input data directly into pmem buffer.

Change-Id: Icd41e7ef406356fc8e6a3e2bdc69e1a172dd72e5