Serialize MountService events
If there are multiple mount events within a short time span (for
instance if auto mount is activated and the usb cable is inserted
and withdrawn in a quick succession) the sd card might not end up in
the correct mount state. This is due to MountService handling mount
events asynchronosly on multiple new threads.
Make MountService handle events in a controlled serial manner, in
one single background thread.

Change-Id: Iff158703112114fb98b1c960563a8aa68bede02d