Update patch set 1
Patch Set 1:

This looks okay, but I'm kind of thinking maybe we should have a "HAVE_SELINUX" include switch for now. We can't break the build of any device when we put this in.

It would be really great if the entire system built without the external/libselinux directory for now so it's easy to "turn off" this feature while we're in experimental mode.

Otherwise if there's a problem, we'll have to revert all the patches which will be a big headache.

Patch-set: 1
Label: Code-Review=0
Label: Verified=0