Battery level based on energy_now/energy_full
If the power supply driver in the kernel doesn't provide with the
'capacity' variable, but provide energy_now and energy_full
variables, the battery level is calculated from these 2 energy_
variables.
This is useful for the x86 ACPI implementation where there's no
capacity variable. It has been tested in a VirtualBox environment.

Change-Id: Ic6afe0a4d3c6ffdabc80114defae7fa403a53d98
Signed-off-by: Daniel Fages <daniel@fages.com>