Update patch set 2
Patch Set 2:

James - would you mind doing a +2 Approved if you're ok with the change (similar for the other ones)? I think JBQ waits for that before he can proceed with the merge. And if you don't have that privilege in the new gerrit yet, please tell him so that he can add that privilege.

Patch-set: 2