Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

I'm not sure that it is correct to replace 1 character string with the space. I.e. if name is "a" it will be replaced with " ". So this will be a bug, again.

Maybe it is better to do like this:
if ((str.length() != 1) && str.startsWith("'") && str.endsWith("'"))

Or what actually the point to remove quotes?
Maybe it easier to use replaceAll?

Patch-set: 2
Reviewer: Gerrit User 1001380 <1001380@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1