Update patch set 2
Patch Set 2:

When the name is a line from supplicant is name='a'
We remove quotes because supplicant adds them.
Root of the problem is the way we get new p2p device notification
from supplicant as a string with space separated fields with format
parameter=value p2p device name can consist of a space. Other solution is to send name from suplicant in hex.

Patch-set: 2