Loop over volume list to find primary for mass storage.
Modify UsbDeviceManager::systemReady() to loop over the volume list
until it finds the primary volume.  Add a boolean method to
StorageVolume that returns true if the volume in question is the
primary storage.

Fixes a problem where the primary storage volume in the list returned
by StorageManager is not the first item.

Change-Id: Iab0927826dfe2629695c7ab5a07d224e02e8fc28
Signed-off-by: Jim Bride <jim.bride@intel.com>
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>