Update patch set 2
Patch Set 2:

Thanks jp for the URL.
Correct me if I'm wrong but UEvent are sent by the kernel ACPI implementation when ACPI events are received from the Hardware/BIOS ; the problem with VirtualBox being that ACPI events are not sent and then correspond UEvents are also never sent.
If my patch is too intrusive into the framework, an other solution might be to have a standalone daemon polling battery information and sending UEvents to the Battery Service (which then wouldn't need any modification).

Dan.

Patch-set: 2
Label: Code-Review=0