Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Following what was said on android-platform. Recommended that I mark this as -1.  This fixes the case for urls such as file:///android_asset/test.html#foo, but breaks on file:///android_asset/test.html?foo=bar.  However, I'm not sure if this is fixable in platform/frameworks/base or if the fix needs to happen in the platform/external/webkit JNI code that connects to this.  Can someone clear this up?

Patch-set: 1