Update patch set 2
Patch Set 2:

It's not intrusive. Passive until property is set on boot.

I don't know how VirtualBox deals with ACPI. But if it is just a configuration issue, then this change is not appropriate. But if VirtualBox+ACPI has an incomplete implementation all the way up and down the sw stack, then this change is ok.

The BIOS normally only sets up the ACPI tables and pokes around the embedded controller that deals with power, then there is a hand-off to the OS.
Do you know where in VirtualBox the ACPI battery event handling is busted? The ASL itself? Does VirtualBox still use QEMU for some hw and maybe acpi handling? Kernel acpi battery driver (apparently not an issue, based on the fact you are saying that ACPI events are not pushed up)?
It would be way cooler to fix the acpi event handling in VirtualBox (asl, kernel drivers), or at least point to a filed bug about it not working.


I'm just trying to make sure you have explored the other likely causes of battery events not working.

Patch-set: 2