Update patch set 2
Patch Set 2:

This indeed fixes the problem. The EditText is no longer editable, which could arguably be what the doc describes.

However, the resulting EditText simply seems to be broken. It looks like a regular EditText, but it does not respond to touch events. When shown next to other regular EditText in a form, this will lead to a terrible user experiencesince there are no other indication that this field is not editable.

I would suggest using an alternative for this use case:
- make the EditText disabled, and enable it when requested.
- use both a TextView and an EditText in your layout and alternate between these by setting one's visibility to GONE.
- change the TextView's attribute to make it become an EditText: inputType, focusable, focusableInTouchMode as well as style attributes like the background.

Patch-set: 2