Update patch set 2
Patch Set 2:

Hi Gilles,

Thank you for the comments.

I know that it looks confusing for the user, but the android:inputType attribute is the XML equivalent of the setInputType() method, and currently doesn't behave the same way. By not fixing this, developers will still call setInputType(InputType.TYPE_NULL) if they insist, but they will be confused, because by using the XML equivalent, will not work, even though it is in the official documentation.

Patch-set: 2