fix NAT detection in SipService
The NAT detection in SipService just checks if the client IP is a
private address. In this case keep alive is enforced. But if you
use a server which also has a private address, keep alive should not
be enabled to save battery.
This patch modifies the NAT detection to also check the server IP.
If both are private, NAT is not detected.

Change-Id: I4b7b420d76fe59851550a26c0e0298266de84e36
Signed-off-by: Andre Valentin <baddream66@googlemail.com>