Start clatd (464xlat) when on a ipv6-only default pdp data context
This adds the class com.android.internal.telephony.Nat464xlatService which
listens for connectivity change broadcast intents to determine when to
start/stop clatd.  When clatd detects dns64/nat64 are in use, it creates
a network interface, "clat".  Nat464xlatService listens for an interfaceAdded
to signal clat is running, and broadcasts the intent
com.android.server.connectivity.nat464xlatservice

LinkProperties was modified to include the type of connection (IP/IPV6/IPV4V6)
as well as being able to handle a different network interface for v4 and v6
traffic.

Tethering was modified to tether to the v4 interface (since tethering is
currently v4 only)

ConnectivityService was modified to choose the interface to route out based on
protocol (used by MMS, HIPRI, and other services when a data connection is
already active).

MobileDataStateTracker listens for the nat464xlatservice broadcast intent to
update the relevant connection's LinkProperties

Change-Id: I2dd8ed9c7f18c654cc20343120f1f4c23559138c
Signed-off-by: Daniel Drown <dan-android@drown.org>